requirements.txt
# PK-MCL 多系统项目依赖
# 核心深度学习框架
torch>=2.2.0

# 科学计算
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0

# 配置文件处理
pyyaml>=6.0

# 进度条
tqdm>=4.66.0
